requirements.txt
asgiref==3.10.0
Brotli==1.1.0
certifi==2025.10.5
charset-normalizer==3.4.4
colorama==0.4.6
Django==5.2.7
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.20.0
fsspec==2025.9.0
idna==3.11
Jinja2==3.1.6
llvmlite==0.45.1
MarkupSafe==3.0.3
more-itertools==10.8.0
mpmath==1.3.0
mutagen==1.47.0
networkx==3.5
numba==0.62.1
numpy==2.3.4
openai-whisper==20250625
pycryptodomex==3.23.0
PyJWT==2.10.1
regex==2025.9.18
requests==2.32.5
setuptools==80.9.0
sqlparse==0.5.3
sympy==1.14.0
tiktoken==0.12.0
torch==2.9.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
yt-dlp==2025.10.14
